requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
transformers==4.41.0
torch==2.3.0
pypdf==4.2.0
python-dotenv==1.0.1
httpx==0.27.0
beautifulsoup4==4.12.3
requests==2.32.2
joblib==1.4.2